2025.06.05-1 updates to scripting, forecast, met-data handling, target temperature handling
We had quite a few fixes come together at the same time, and consequently have a new WTMP build (2025.06.05-1) that requires updated studies.  The studies have some changes to configuration files in the forecast folder, and scripting updates, but also have new tags in the simulation group files that necessitates rebuilding the met data sets (and boundary condition sets).  We have done that for the example forecasts in the GIT studies.  

** If you want/need to preserve current forecast simulation groups, let us know and we can find a workaround to updating your WAT studies.  You will still need to rebuild the met data sets/boundary condition data sets.

This WTMP build / study update addresses the following issues, which I will either mark as complete or will re-assigned as testing if further verification is needed:

•	240 (gate configuration initialization) – Fixed/Complete

•	234 (American forecast temp target dates incorrect) 

•	235 (Temperature target matrix defaulting to weekly)  

•	157 (L3MTO) – done/testing

•	233 – American seasonal ave air temp – fixed/complete

•	212 – American met cloudiness missing – fixed/complete

•	158 – More recent met – fixed/complete

•	215 – Linked Shasta W2 – ResSim model alt – done/testing (but no Shasta 5.0)

•	237 – American wind data for forecast data download – complete